Moved preprocessing to its own package
