+++ preprocess_comparison/compare_preprocess.py
@@ -2,9 +2,9 @@ from absl import app, flags
 
 from preprocess_comparison.comparison_utils import plot_weights, available_metrics
 from preprocess_comparison.load import dataset_modules, get_client_weights
-from shared.truncate import truncate
-from shared.lp import lp
-from shared.utils import is_valid_solution, maximal_weight_proportion
+from shared.preprocess.truncate import truncate
+from shared.preprocess.lp import lp
+from shared.preprocess.utils import is_valid_solution, maximal_weight_proportion
 from google_tff_research.utils import utils_impl
 
 from pprint import PrettyPrinter
